+++ ext/stub/issue2656.zep.c
@@ -64,7 +64,7 @@ PHP_METHOD(Stub_Issue2656, setContainer)
 		Z_PARAM_ZVAL(container)
 	ZEND_PARSE_PARAMETERS_END();
 	zephir_fetch_params_without_memory_grow(1, 0, &container);
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 65, container);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 66, container);
 }
 
 /**
@@ -245,7 +245,7 @@ PHP_METHOD(Stub_Issue2656, fetchThroughProperty)
 	ZEPHIR_INIT_VAR(&value);
 	ZVAL_NULL(&value);
 	ZEPHIR_OBS_NVAR(&value);
-	zephir_read_property_cached(&_0, this_ptr, _zephir_prop_0, 65, PH_NOISY_CC | PH_READONLY);
+	zephir_read_property_cached(&_0, this_ptr, _zephir_prop_0, 66, PH_NOISY_CC | PH_READONLY);
 	if (zephir_array_isset_string_fetch(&value, &_0, SL("cache.key"), 0)) {
 		RETURN_CCTOR(&value);
 	}