+++ ext/stub/properties/readonlyproperties.zep.c
@@ -98,11 +98,11 @@ PHP_METHOD(Stub_Properties_ReadonlyProperties, __construct)
 
 	ZVAL_UNDEF(&_0);
 	ZVAL_LONG(&_0, 42);
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 100, &_0);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 101, &_0);
 	ZEPHIR_INIT_VAR(&_1);
 	ZEPHIR_INIT_NVAR(&_1);
 	ZVAL_STRING(&_1, "hello");
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_1, 101, &_1);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_1, 102, &_1);
 	ZEPHIR_INIT_VAR(&_2);
 	zephir_create_array(&_2, 2, 0);
 	ZEPHIR_INIT_NVAR(&_1);
@@ -111,11 +111,11 @@ PHP_METHOD(Stub_Properties_ReadonlyProperties, __construct)
 	ZEPHIR_INIT_NVAR(&_1);
 	ZVAL_STRING(&_1, "b");
 	zephir_array_fast_append(&_2, &_1);
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_2, 102, &_2);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_2, 103, &_2);
 	if (1) {
-		zephir_update_property_zval_cached(this_ptr, _zephir_prop_3, 103, &__$true);
+		zephir_update_property_zval_cached(this_ptr, _zephir_prop_3, 104, &__$true);
 	} else {
-		zephir_update_property_zval_cached(this_ptr, _zephir_prop_3, 103, &__$false);
+		zephir_update_property_zval_cached(this_ptr, _zephir_prop_3, 104, &__$false);
 	}
 	ZEPHIR_INIT_NVAR(&_1);
 	object_init_ex(&_1, stub_properties_publicproperties_ce);
@@ -124,7 +124,7 @@ PHP_METHOD(Stub_Properties_ReadonlyProperties, __construct)
 		zephir_check_call_status();
 	}
 
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_4, 104, &_1);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_4, 105, &_1);
 	ZEPHIR_MM_RESTORE();
 }